backend/src/main/java/com/codurance/codurawise/lambdas/AddResource.java
package com.codurance.codurawise.lambdas;

import com.amazonaws.services.lambda.runtime.Context;
import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyRequestEvent;
import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyResponseEvent;
import com.codurance.codurawise.domain.services.ResourceService;

public class AddResource {
  public AddResource(ResourceService resourceService) {

  }

  public APIGatewayProxyResponseEvent handleRequest(APIGatewayProxyRequestEvent requestEvent, Context context) {
    return null;
  }
}
